# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Practica Parcial 1'","description":"'LCIV Practica de parcial con uso de Docker y Docker Compose'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCIV Practica de parcial con uso de Docker y Docker Compose'"}],"paths":{"/clients":{"post":{"tags":["user-controller"],"operationId":"setCliente","parameters":[{"name":"userRequestDTO","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/weather/locations":{"get":{"tags":["weather-controller"],"operationId":"getLocations","parameters":[{"name":"client_id","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"secret","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Locations"}}}}}}}},"/weather/locations/datetime":{"get":{"tags":["weather-controller"],"operationId":"getLocationsByDate","parameters":[{"name":"client_id","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"secret","in":"header","required":true,"schema":{"type":"string"}},{"name":"idLocation","in":"header","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeatherResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UserRequestDTO":{"type":"object","properties":{"email":{"type":"string"},"temperature_unit":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"client_id":{"type":"integer","format":"int64"},"secret":{"type":"string"}}},"Locations":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"latitude":{"type":"string"},"longitude":{"type":"string"}}},"AirQualityDTO":{"type":"object","properties":{"index":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"CloudinessDTO":{"type":"object","properties":{"index":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"LocationDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"TemperatureDTO":{"type":"object","properties":{"value":{"type":"integer","format":"int32"},"unit":{"type":"string"}}},"WeatherResponseDTO":{"type":"object","properties":{"location":{"$ref":"#/components/schemas/LocationDTO"},"temperature":{"$ref":"#/components/schemas/TemperatureDTO"},"wind":{"type":"string"},"air_quality":{"$ref":"#/components/schemas/AirQualityDTO"},"cloudiness":{"$ref":"#/components/schemas/CloudinessDTO"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}