# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CloudinessDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class CloudinessDTO {
    private int index;
    private String description;
}