# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/UserRequestDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import jakarta.validation.constraints.Email;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class UserRequestDTO {
    @Email
    private String email;
    private String temperature_unit;
}