# src/main/java/ar/edu/utn/frc/tup/lciii/models/Locations.java
package ar.edu.utn.frc.tup.lciii.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Locations {
    private Long id;
    private String name;
    private String latitude;
    private String longitude;
}